Format
Hey guys here is what I have so far for formatting. Its not as fancy as I wanted to do, but the other way got too complicate. I might still try to play around with a fancier parallax if I have some more time this weekend or depending on how much time we have next week.